+++ API/Business/Trolley/DTOs/TrolleyReadDTOForUser.cs
@@ -7,7 +7,7 @@ namespace Business.Trolley.DTOs
     public class TrolleyReadDTOForUser
     {
         public int UserId { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
         public IEnumerable<CatalogueProductReadDTO_View> CatalogueProducts { get; set; }
     }
 }
